Reduce linewidth type to uint8, since that's enough, as it doesn't need to be sacled anyway.

+++ internal/game/lobby.go
@@ -117,9 +117,9 @@ func (lobby *Lobby) HandleEvent(eventType string, payload []byte, player *Player
 
 			// In case the line is too big, we overwrite the data of the event.
 			// This will prevent clients from lagging due to too thick lines.
-			if line.Data.LineWidth > float32(MaxBrushSize) {
+			if line.Data.LineWidth > MaxBrushSize {
 				line.Data.LineWidth = MaxBrushSize
-			} else if line.Data.LineWidth < float32(MinBrushSize) {
+			} else if line.Data.LineWidth < MinBrushSize {
 				line.Data.LineWidth = MinBrushSize
 			}
